Abstracting changelog to a testable module
