+++ src/utils/constants/index.ts
@@ -1,0 +1,7 @@
+import locations from "./locations";
+import changelog from "./changelog";
+
+export {
+  locations,
+  changelog
+}
