+++ src/utils/index.ts
@@ -1,5 +1,9 @@
 import * as network from "./network";
+import {
+  locations
+ } from "./constants";
 
 export {
-  network
+  network,
+  locations
 };
